Streamline bundler reports and metrics
